pyproject.toml
[project]
name = "tch-mentormatching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.12.13",
    "asyncio>=3.4.3",
    "faiss-cpu==1.8.0",
    "gpt4all==2.7.0",
    "gradio==4.44.0",
    "gradio-client==1.3.0",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.1.21",
    "openai>=1.86.0",
    "pandas==2.2.2",
    "pypdf==4.2.0",
    "pypdf2==3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv==1.0.1",
    "tenacity==8.4.1",
    "tqdm==4.66.4",
]
